Implement Nav3
Implemented Nav3 on top of existing Navigation. Also upgraded the target SDK to 36 (Nav3 required API 36) along with the gradle plugin to the latest versions.
