+++ gradle/libs.versions.toml
@@ -8,7 +8,7 @@ accompanist = "0.36.0"
 work = "2.10.0"
 exoplayer = "2.19.1"
 paging = "3.3.5"
-androidGradlePlugin = "8.5.2"
+androidGradlePlugin = "8.10.0"
 sentry = "5.3.0"
 ksp = "2.0.0-1.0.23"
 composeBom = "2025.01.00"
@@ -45,11 +45,14 @@ turbine = "1.1.0"
 vicoCompose = "2.0.0-alpha.22"
 composeRatingbar = "1.3.12"
 loggerAndroid = "1.0.0"
-compileSdk = "35"
-targetSdk = "35"
+compileSdk = "36"
+targetSdk = "36"
 minSdk = "22"
 paletteKtx = "1.0.0"
 chucker = "4.1.0"
+nav3Core = "1.0.0-alpha02"
+nav3Material = "1.0.0-SNAPSHOT"
+nav3Lifecycle = "1.0.0-alpha01"
 
 [libraries]
 androidx-appcompat = { module = "androidx.appcompat:appcompat", version.ref = "appcompat" }
@@ -147,6 +150,11 @@ androidx-palette-ktx = { group = "androidx.palette", name = "palette-ktx", versi
 
 chucker = { module = "com.github.chuckerteam.chucker:library", version.ref = "chucker" }
 chucker-noop = { module = "com.github.chuckerteam.chucker:library-no-op", version.ref = "chucker" }
+
+androidx-navigation3-runtime = { module = "androidx.navigation3:navigation3-runtime", version.ref = "nav3Core" }
+androidx-navigation3-ui = { module = "androidx.navigation3:navigation3-ui", version.ref = "nav3Core" }
+androidx-lifecycle-viewmodel-navigation3 = { module = "androidx.lifecycle:lifecycle-viewmodel-navigation3", version.ref = "nav3Lifecycle" }
+
 
 [plugins]
 android-application = { id = "com.android.application", version.ref = "androidGradlePlugin" }
